+++ Ddugyesangi/Views/Part/PartEditView.swift
@@ -28,7 +28,7 @@ struct PartEditView: View {
     }
     
     var body: some View {
-        NavigationView {
+        NavigationStack {
             ZStack {
                 ThemedBackgroundView()
                     .ignoresSafeArea()